refactor(queries): fold the dispatcher method groups back into four files (15 -> 8)
Every symbol in internal/queries hangs off one type, *QueryDispatcher, and
Go methods cannot live in another package — so subpackages here would mean
either exporting the two mutexes and their ordering invariant or duplicating
state. The reference graph confirms the shape: 12 of 15 files reference
dispatcher.go or types.go and nothing forms an independent cluster.

What was actually splittable was already split too far. Commits 58264122 and
54163831 sliced the command/query/result method groups across 10 files with
no seam between them — dispatcher_results_{store,wait,cleanup} all mutate
queryResults and pendingQueries under the same mu, and _cleanup even calls
back into _store's neighbours. Folded them back:

  dispatcher_commands.go  <- +_observe.go +_wait.go
  dispatcher_queries.go   <- +_poll.go +_expire.go
  dispatcher_results.go   <- +_store.go +_wait.go +_cleanup.go

Code moved verbatim (verified: the multiset of non-comment code lines is
identical to HEAD); the import path is unchanged, so none of the ~60
consumers of internal/queries move. Largest resulting file is 372 lines.

dispatcher_trace.go stays separate but is deliberately NOT extracted into a
subpackage: ensureTraceContextLocked/hasTraceStageLocked/appendTraceEventLocked
are all called with resultsMu held by the command path.

internal/queries drops from 15 to 8 source files and out of
.folder-size-baseline.json entirely.
