+++ docs/features/feature/query-service/index.md
@@ -8,12 +8,22 @@ last_reviewed: 2026-07-26
 code_paths:
   - internal/mcp/response_json.go
   - internal/mcp/response_builders.go
+  - internal/queries/dispatcher.go
+  - internal/queries/dispatcher_commands.go
+  - internal/queries/dispatcher_queries.go
+  - internal/queries/dispatcher_results.go
+  - internal/queries/dispatcher_trace.go
+  - internal/queries/types.go
   - src/types/index.ts
   - src/types/messages.ts
   - src/types/runtime-messages.ts
   - src/types/runtime/queries.ts
 test_paths:
   - internal/mcp/response_test.go
+  - internal/queries/dispatcher_test.go
+  - internal/queries/commands_test.go
+  - internal/queries/command_trace_test.go
+  - internal/queries/expire_signal_test.go
 last_verified_version: 0.7.12
 last_verified_date: 2026-03-05
 ---
